Added colors to shared constants. _yield function. Added sse to libc. Added mandelbrot user program.

+++ Userland/libc/syscall.c
@@ -59,7 +59,7 @@ void setForeground(uint8_t red, uint8_t green, uint8_t blue) {
 
 void drawBitmap(uint16_t x, uint16_t y, uint16_t width, uint16_t height,
                 Color bitmap[][width]) {
-	syscall(DRAWBITMAP, x, y, width, height, bitmap);
+	syscall(DRAWBITMAP, x, y, width, height, (uint64_t)bitmap);
 }
 
 WindowInfo getWindowInfo() {
